requirements.txt
openai
langchain
langchain-openai
langchain-community
pandas
numpy
notebook
faiss-cpu
python-dotenv
streamlit
prefect
pypdf
unstructured
networkx
openpyxl
tiktoken
streamlit-chat
audio_recorder_streamlit
streamlit-float